I'm kinda bored.
A lot of things - big and small. I didn't make commits, because I was
experimenting with a lot of stuff.
1. I finished the "user" part of the application - it's possible to
actually run it now like a normal human being. Wow.
2. I understand Gradle better (like 1% overall, compared to 0.01%
before) - got the build file cleaned up a bit.
3. Oh, right. New task 'executable' to compile the whole thing to a
sorta independent executable. The task adds a custom, jlinked JRE,
launch4j's .exe and sample automata to a single folder.* Kinda cool.
4. Added sample automatas.

----------------------------------------------------------------------

On that note, I'm kinda bored of this project.

What's left:
1. More flexible / less repetetive functions might never be implemented
(for example, right now, the 'is' is part of the 'sum' function and
every function returns boolean. Functions also can't overlap - something
like 'north is [state]' and 'north west is [state]' is not possible - a
task perfect for some kind of a FSM).
2. Speed up the simulation. This project might actually be interesting
to try and implement 'hashlife' in.
For evaluating expressions, maybe something like a bytecode VM? That
last one is probably overkill, it would be better after porting to C/C++
or even Rust (as a learning exercise)?
3. Improve the language. When trying to write a simple fluid simulation,
I had to repeat a lot of cases (ex. 5 lines with 'FULL:' for example).
So, a general 'switch-case' structure? A lot of writing, basically. It's
interesting, because the language might be the greatest barrier in
creating some interesting cellular automata.

(3) is similar to (1) in a way, that we can increase the expressiveness
of this language by better implementing function definitions.
Instead of 4 times 'north is GOES_DOWN and random 1 in n chance ->
[state]'
Adding better functions can change it to: 'north is GOES_DOWN -> choose
GOES_UP, GOES_DOWN, GOES_LEFT, GOES_RIGHT', where 'choose' is of type
State.


What's left, but might actually be implemented, because I'm not as tired
of this project as I thought:
1. Toroidal-ness.
2. Setting (maximum) simulation speed. It ain't that fast and I don't
know no 'hashlife'. Git outta here!
3. Remove requirement for specyfing neighborhood type (like, it's not
needed if the CA is not using it - random walk for example).
4. Change the order of color definitions - originally there, because of
code reuse (laziness) and that it kinda makes sense that states are
differentiated by color first (what you see), then the state it belongs
to (what it actually is). It breaks down on something like Langton's
Ant, where multiple states are red. So, yeah, one-to-many: [state] is
[color].
5. Custom color, represented by a leading '#', then 6 hex digits. Should
be added after implementing (4), because it'll look bad otherwise (ex.
'#00FF00 is ON' vs 'ON is #00FF00').
6. Parameters (and a system to handle them) to configure all these.
7. Images instead of colors. The framework is there (State uses
BufferedImage by default), but we might have to decouple creating states
from parsing? (Suppose we specify the default image folder in 'Defualts'
and reference images only by filename. We can do it while parsing, but
it'll lead to some 'half-parameters' which are only valid inside the
Parser object and would need to be removed after.)
8. And, as always, finish pretty printing for error reporting. (^^^^^ to
underline errors).

-----------------------------------------------------------------------

*I experimented with two ways to bundle an executable.
1. Official - Making an installer. Why I didn't go for it? I don't feel
like this manlet of an application should require a whole installation
process.

2. Hacky - Something called 'sfx' (self-extracting archive). I actually
ended up making it work (manually), but, because it seems to be an
untrustoworthy format
(https://en.wikipedia.org/wiki/Self-extracting_archive#Disadvantages ),
Windows always shows its password prompt, which is really annoying. So,
I scrapped it, but the idea for both a JRE and a program bundled into a
single executable (without it being obvious it uses a virtual machine)
is interesting.
Also, the fact that the jlink Gradle plugin also supports jpackage and
creating the whole "image" structure, which I used for sfx, is a welcome
surprise. (I'm not sure if this 'image structure' is the jpackage's or
the plugin's doing.)